refactor: remove pass-through methods
