+++ src/utils/files/index.ts
@@ -1,5 +1,6 @@
 export { WorkspaceFS } from './workspaceFS';
 export { AbsoluteFS } from './absoluteFS';
 export { StorageFS, GlobalStorageFS } from './storageFS';
+export { RelativeFS } from './relativeFS';
 export * from './pastedImageUtils';
 export * from './baseFileUtils';
