+++ src/utils/files/storageFS.ts
@@ -4,6 +4,9 @@ import * as path from 'path';
 // Third-party imports
 import * as vscode from 'vscode';
 
+// Local imports - fs
+import { RelativeFS } from './relativeFS';
+
 // Local imports - log
 import * as logger from '@logger/logUtils';
 
@@ -14,7 +17,7 @@ logger.initialize(CHANNEL);
  * StorageFS provides a unified interface for VS Code extension storage operations.
  * Supports both workspace storage (per-workspace) and global storage (shared across workspaces).
  */
-export class StorageFS {
+export class StorageFS extends RelativeFS {
   private static context: vscode.ExtensionContext | null = null;
 
   /**
@@ -48,49 +51,8 @@ export class StorageFS {
     return this.context.globalStorageUri.fsPath;
   }
 
-  /**
-   * Get the full path for a relative storage path
-   */
-  public static fullPath(relativePath: string): string {
-    return path.join(this.getPath(), relativePath);
-  }
-
-  /**
-   * Check if a file or directory exists
-   */
-  public static async exists(relativePath: string): Promise<boolean> {
-    try {
-      const fullPath = this.fullPath(relativePath);
-      const uri = vscode.Uri.file(fullPath);
-      await vscode.workspace.fs.stat(uri);
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * Read a file as UTF-8 text
-   */
-  public static async read(relativePath: string): Promise<string> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    const content = await vscode.workspace.fs.readFile(uri);
-    return Buffer.from(content).toString('utf-8');
-  }
-
-  /**
-   * Write content to a file (text or binary)
-   */
-  public static async write(
-    relativePath: string,
-    content: string | Uint8Array,
-  ): Promise<void> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    const data =
-      typeof content === 'string' ? Buffer.from(content, 'utf-8') : content;
-    await vscode.workspace.fs.writeFile(uri, data);
+  protected static override getBasePath(): string {
+    return this.getPath();
   }
 
   /**
@@ -100,9 +62,7 @@ export class StorageFS {
     relativePath: string,
     options?: { recursive?: boolean; useTrash?: boolean },
   ): Promise<void> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    await vscode.workspace.fs.delete(uri, options);
+    await super.delete(relativePath, options);
     logger.debug(CHANNEL, `Deleted: ${relativePath}`);
   }
 
@@ -110,49 +70,8 @@ export class StorageFS {
    * Create a directory
    */
   public static async createDir(relativePath: string): Promise<void> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    await vscode.workspace.fs.createDirectory(uri);
+    await super.createDir(relativePath);
     logger.debug(CHANNEL, `Created directory: ${relativePath}`);
-  }
-
-  /**
-   * Ensure a directory exists, creating it if necessary
-   */
-  public static async ensureDir(relativePath: string): Promise<void> {
-    try {
-      const exists = await this.exists(relativePath);
-      if (!exists) {
-        await this.createDir(relativePath);
-      }
-    } catch (err) {
-      // If error is because directory already exists, ignore it
-      if (err instanceof vscode.FileSystemError && err.code === 'FileExists') {
-        return;
-      }
-      // Re-throw other errors (permission denied, disk full, etc.)
-      throw err;
-    }
-  }
-
-  /**
-   * Read directory contents
-   */
-  public static async readDir(
-    relativePath: string,
-  ): Promise<[string, vscode.FileType][]> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    return await vscode.workspace.fs.readDirectory(uri);
-  }
-
-  /**
-   * Get file stats
-   */
-  public static async stat(relativePath: string): Promise<vscode.FileStat> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    return await vscode.workspace.fs.stat(uri);
   }
 
   /**
@@ -163,9 +82,7 @@ export class StorageFS {
     destination: string,
     options?: { overwrite?: boolean },
   ): Promise<void> {
-    const sourceUri = vscode.Uri.file(this.fullPath(source));
-    const destUri = vscode.Uri.file(this.fullPath(destination));
-    await vscode.workspace.fs.copy(sourceUri, destUri, options);
+    await super.copy(source, destination, options);
     logger.debug(
       CHANNEL,
       `Copied: source=${source} to destination=${destination}`,
@@ -180,124 +97,22 @@ export class StorageFS {
     newPath: string,
     options?: { overwrite?: boolean },
   ): Promise<void> {
-    const oldUri = vscode.Uri.file(this.fullPath(oldPath));
-    const newUri = vscode.Uri.file(this.fullPath(newPath));
-    await vscode.workspace.fs.rename(oldUri, newUri, options);
+    await super.rename(oldPath, newPath, options);
     logger.debug(CHANNEL, `Renamed: ${oldPath} to ${newPath}`);
   }
 
-  /**
-   * Clean up old files in a directory based on age
-   */
-  public static async cleanupOldFiles(
-    relativePath: string,
-    maxAgeMs: number,
-  ): Promise<void> {
-    try {
-      const entries = await this.readDir(relativePath);
-      const now = Date.now();
-
-      for (const [name, type] of entries) {
-        if (type !== vscode.FileType.File) {
-          continue;
-        }
-
-        const filePath = path.join(relativePath, name);
-        try {
-          const stats = await this.stat(filePath);
-          if (now - stats.mtime > maxAgeMs) {
-            await this.delete(filePath);
-            logger.debug(CHANNEL, `Deleted old file: ${filePath}`);
-          }
-        } catch (err) {
-          logger.warn(
-            CHANNEL,
-            `Error checking file age ${filePath}: ${err instanceof Error ? err.message : String(err)}`,
-          );
-        }
-      }
-    } catch (err) {
-      logger.warn(
-        CHANNEL,
-        `Error cleaning directory ${relativePath}: ${err instanceof Error ? err.message : String(err)}`,
-      );
-    }
-  }
-
-  /**
-   * Check if path is a directory
-   */
-  public static async isDir(relativePath: string): Promise<boolean> {
-    try {
-      const stats = await this.stat(relativePath);
-      return stats.type === vscode.FileType.Directory;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * Check if path is a file
-   */
-  public static async isFile(relativePath: string): Promise<boolean> {
-    try {
-      const stats = await this.stat(relativePath);
-      return stats.type === vscode.FileType.File;
-    } catch {
-      return false;
-    }
-  }
+  // Inherit cleanupOldFiles, isDir and isFile from RelativeFS
 }
 
 /**
  * GlobalStorageFS provides operations for global storage (shared across workspaces).
  * Extends StorageFS but uses global storage paths instead of workspace storage.
  */
-export class GlobalStorageFS {
-  /**
-   * Get the full path for a relative global storage path
-   */
-  public static fullPath(relativePath: string): string {
-    return path.join(StorageFS.getGlobalPath(), relativePath);
+export class GlobalStorageFS extends RelativeFS {
+  protected static override getBasePath(): string {
+    return StorageFS.getGlobalPath();
   }
-
-  /**
-   * Check if a file or directory exists in global storage
-   */
-  public static async exists(relativePath: string): Promise<boolean> {
-    try {
-      const fullPath = this.fullPath(relativePath);
-      const uri = vscode.Uri.file(fullPath);
-      await vscode.workspace.fs.stat(uri);
-      return true;
-    } catch {
-      return false;
-    }
-  }
-
-  /**
-   * Read a file from global storage as UTF-8 text
-   */
-  public static async read(relativePath: string): Promise<string> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    const content = await vscode.workspace.fs.readFile(uri);
-    return Buffer.from(content).toString('utf-8');
-  }
-
-  /**
-   * Write content to a file in global storage (text or binary)
-   */
-  public static async write(
-    relativePath: string,
-    content: string | Uint8Array,
-  ): Promise<void> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    const data =
-      typeof content === 'string' ? Buffer.from(content, 'utf-8') : content;
-    await vscode.workspace.fs.writeFile(uri, data);
-  }
+  // Inherit path helpers and basic operations from RelativeFS
 
   /**
    * Delete a file or directory from global storage
@@ -306,9 +121,7 @@ export class GlobalStorageFS {
     relativePath: string,
     options?: { recursive?: boolean; useTrash?: boolean },
   ): Promise<void> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    await vscode.workspace.fs.delete(uri, options);
+    await super.delete(relativePath, options);
     logger.debug(CHANNEL, `Deleted from global storage: ${relativePath}`);
   }
 
@@ -316,53 +129,14 @@ export class GlobalStorageFS {
    * Create a directory in global storage
    */
   public static async createDir(relativePath: string): Promise<void> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    await vscode.workspace.fs.createDirectory(uri);
+    await super.createDir(relativePath);
     logger.debug(
       CHANNEL,
       `Created directory in global storage: ${relativePath}`,
     );
   }
 
-  /**
-   * Ensure a directory exists in global storage, creating it if necessary
-   */
-  public static async ensureDir(relativePath: string): Promise<void> {
-    try {
-      const exists = await this.exists(relativePath);
-      if (!exists) {
-        await this.createDir(relativePath);
-      }
-    } catch (err) {
-      // If error is because directory already exists, ignore it
-      if (err instanceof vscode.FileSystemError && err.code === 'FileExists') {
-        return;
-      }
-      // Re-throw other errors (permission denied, disk full, etc.)
-      throw err;
-    }
-  }
-
-  /**
-   * Read directory contents from global storage
-   */
-  public static async readDir(
-    relativePath: string,
-  ): Promise<[string, vscode.FileType][]> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    return await vscode.workspace.fs.readDirectory(uri);
-  }
-
-  /**
-   * Get file stats from global storage
-   */
-  public static async stat(relativePath: string): Promise<vscode.FileStat> {
-    const fullPath = this.fullPath(relativePath);
-    const uri = vscode.Uri.file(fullPath);
-    return await vscode.workspace.fs.stat(uri);
-  }
+  // Inherit ensureDir, readDir and stat from RelativeFS
 
   /**
    * Copy a file or directory in global storage
@@ -372,9 +146,7 @@ export class GlobalStorageFS {
     destination: string,
     options?: { overwrite?: boolean },
   ): Promise<void> {
-    const sourceUri = vscode.Uri.file(this.fullPath(source));
-    const destUri = vscode.Uri.file(this.fullPath(destination));
-    await vscode.workspace.fs.copy(sourceUri, destUri, options);
+    await super.copy(source, destination, options);
     logger.debug(
       CHANNEL,
       `Copied in global storage: source=${source} to destination=${destination}`,
@@ -389,9 +161,7 @@ export class GlobalStorageFS {
     newPath: string,
     options?: { overwrite?: boolean },
   ): Promise<void> {
-    const oldUri = vscode.Uri.file(this.fullPath(oldPath));
-    const newUri = vscode.Uri.file(this.fullPath(newPath));
-    await vscode.workspace.fs.rename(oldUri, newUri, options);
+    await super.rename(oldPath, newPath, options);
     logger.debug(
       CHANNEL,
       `Renamed in global storage: ${oldPath} to ${newPath}`,
